requirements.txt
# Basic Web Lib
pymysql
mysqlclient
Django
SQLAlchemy
sympy
djangorestframework
requests
django-rest-knox
django-cors-headers

# package use for store file into ASW s3
django-storages
boto3

# Deal With Image
Pillow


# Packages for Deep Learning Model
matplotlib>=3.2.2
numpy>=1.18.5
# opencv-python>=4.1.1
opencv-python-headless
PyYAML>=5.3.1
scipy>=1.4.1
torch>=1.7.0,!=1.12.0
torchvision>=0.8.1,!=0.13.0
tqdm>=4.41.0
protobuf<4.21.3

# Logging -------------------------------------
tensorboard>=2.4.1
# wandb

# Plotting ------------------------------------
pandas>=1.1.4
seaborn>=0.11.0

# Extras --------------------------------------
ipython  # interactive notebook
psutil  # system utilization
thop  # FLOPs computation